refactor(agents): isolate vendored Vercel stream reducer
