+++ src/phoenix/server/api/routers/agents.py
@@ -141,8 +141,10 @@ from phoenix.server.agents.types import (
 )
 from phoenix.server.agents.ui_message_stream import (
     AgentErrorChunk,
-    create_streaming_ui_message_state,
     iter_chunks_with_error_parts,
+)
+from phoenix.server.agents.vercel_ui_message_stream import (
+    create_streaming_ui_message_state,
     reduce_ui_message_chunk,
 )
 from phoenix.server.api.helpers.agent_sessions import TURN_LOCK_STALENESS, is_turn_active
